Fix replace images
